// src/app/api/models/call-list-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CallListResponse {
  data: Array<{
'id': string;
'callId': string;
'agentId': string | null;
'callType': string | null;
'direction': string | null;
'fromNumber': string | null;
'toNumber': string | null;
'callStatus': string | null;
'disconnectionReason': string | null;
'startedAt': string | null;
'endedAt': string | null;
'durationMs': number | null;
'userSentiment': string | null;
'callSummary': string | null;
'transcript': string | null;
'recordingUrl': string | null;
'isRead': boolean;
'status': 'active' | 'suspended';
'createdAt': string;
'updatedAt': string;
'deletedAt': string | null;
}>;
  limit: number;
  page: number;
  total: number;
  totalPages: number;
}